Move monthly emission changes to transportation calculations file and reducer
