+++ client/src/app/calculations.ts
@@ -5,40 +5,18 @@ import {
   EereDefaultData,
 } from 'app/redux/reducers/geography';
 import { EereTextInputFieldName } from 'app/redux/reducers/eere';
-// config
-import { percentHybridEVMilesDrivenOnElectricity } from 'app/config';
 // calculations
 import type {
-  MonthlyVMTPerVehicle,
   DailyStats,
   HourlyEVChargingPercentages,
-  VehiclesDisplaced,
   MonthlyEVEnergyUsageGW,
   MonthlyDailyEVEnergyUsage,
-  MonthlyEmissionRates,
+  MonthlyEmissionChanges,
 } from 'app/calculations/transportation';
 
 const pollutants = ['CO2', 'NOX', 'SO2', 'PM25', 'VOCs', 'NH3'] as const;
 
 type Pollutant = typeof pollutants[number];
-
-type ExpandedVehicleType =
-  | 'batteryEVCars'
-  | 'hybridEVCars'
-  | 'batteryEVTrucks'
-  | 'hybridEVTrucks'
-  | 'transitBusesDiesel'
-  | 'transitBusesCNG'
-  | 'transitBusesGasoline'
-  | 'schoolBuses';
-
-type GeneralVehicleType =
-  | 'cars'
-  | 'trucks'
-  | 'transitBusesDiesel'
-  | 'transitBusesCNG'
-  | 'transitBusesGasoline'
-  | 'schoolBuses';
 
 /**
  * Totals the energy usage from each EV type for all months in the year to a
@@ -56,105 +34,16 @@ export function calculateTotalYearlyEVEnergyUsage(
 }
 
 /**
- * Monthly emission changes by EV type.
- *
- * Excel: Top half of the "Emission Changes" data from "Table 7: Calculated
- * changes for the transportation sector" table in the "Library" sheet
- * (G316:R336).
- */
-function calculateMonthlyEmissionChangesByEVType(options: {
-  monthlyVMTPerVehicle: MonthlyVMTPerVehicle;
-  vehiclesDisplaced: VehiclesDisplaced;
-  monthlyEmissionRates: MonthlyEmissionRates;
-}) {
-  const { monthlyVMTPerVehicle, vehiclesDisplaced, monthlyEmissionRates } =
-    options;
-
-  const result: {
-    [month: number]: {
-      [vehicleType in ExpandedVehicleType]: {
-        [pollutant in Pollutant]: number;
-      };
-    };
-  } = {};
-
-  Object.entries(monthlyEmissionRates).forEach(([key, data]) => {
-    const month = Number(key);
-
-    result[month] ??= {
-      batteryEVCars: { CO2: 0, NOX: 0, SO2: 0, PM25: 0, VOCs: 0, NH3: 0 },
-      hybridEVCars: { CO2: 0, NOX: 0, SO2: 0, PM25: 0, VOCs: 0, NH3: 0 },
-      batteryEVTrucks: { CO2: 0, NOX: 0, SO2: 0, PM25: 0, VOCs: 0, NH3: 0 },
-      hybridEVTrucks: { CO2: 0, NOX: 0, SO2: 0, PM25: 0, VOCs: 0, NH3: 0 },
-      transitBusesDiesel: { CO2: 0, NOX: 0, SO2: 0, PM25: 0, VOCs: 0, NH3: 0 },
-      transitBusesCNG: { CO2: 0, NOX: 0, SO2: 0, PM25: 0, VOCs: 0, NH3: 0 },
-      transitBusesGasoline: { CO2: 0, NOX: 0, SO2: 0, PM25: 0, VOCs: 0, NH3: 0 }, // prettier-ignore
-      schoolBuses: { CO2: 0, NOX: 0, SO2: 0, PM25: 0, VOCs: 0, NH3: 0 },
-    };
-
-    pollutants.forEach((pollutant) => {
-      result[month].batteryEVCars[pollutant] =
-        data.cars[pollutant] *
-        monthlyVMTPerVehicle[month].cars *
-        vehiclesDisplaced.batteryEVCars;
-
-      result[month].hybridEVCars[pollutant] =
-        data.cars[pollutant] *
-        monthlyVMTPerVehicle[month].cars *
-        vehiclesDisplaced.hybridEVCars *
-        (percentHybridEVMilesDrivenOnElectricity / 100);
-
-      result[month].batteryEVTrucks[pollutant] =
-        data.trucks[pollutant] *
-        monthlyVMTPerVehicle[month].trucks *
-        vehiclesDisplaced.batteryEVTrucks;
-
-      result[month].hybridEVTrucks[pollutant] =
-        data.trucks[pollutant] *
-        monthlyVMTPerVehicle[month].trucks *
-        vehiclesDisplaced.hybridEVTrucks *
-        (percentHybridEVMilesDrivenOnElectricity / 100);
-
-      result[month].transitBusesDiesel[pollutant] =
-        data.transitBusesDiesel[pollutant] *
-        monthlyVMTPerVehicle[month].transitBusesDiesel *
-        vehiclesDisplaced.transitBusesDiesel;
-
-      result[month].transitBusesCNG[pollutant] =
-        data.transitBusesCNG[pollutant] *
-        monthlyVMTPerVehicle[month].transitBusesCNG *
-        vehiclesDisplaced.transitBusesCNG;
-
-      result[month].transitBusesGasoline[pollutant] =
-        data.transitBusesGasoline[pollutant] *
-        monthlyVMTPerVehicle[month].transitBusesGasoline *
-        vehiclesDisplaced.transitBusesGasoline;
-
-      result[month].schoolBuses[pollutant] =
-        data.schoolBuses[pollutant] *
-        monthlyVMTPerVehicle[month].schoolBuses *
-        vehiclesDisplaced.schoolBuses;
-    });
-  });
-
-  return result;
-}
-
-/**
  * Totals monthly emission changes from each EV type.
  *
  * Excel: Bottom half of the "Emission Changes" data from "Table 7: Calculated
  * changes for the transportation sector" table in the "Library" sheet
  * (G336:R394).
  */
-function calculateTotalMonthlyEmissionChanges(monthlyEmissionChangesByEVType: {
-  [month: number]: {
-    [vehicleType in ExpandedVehicleType]: {
-      [pollutant in Pollutant]: number;
-    };
-  };
-}) {
-  const result = Object.entries(monthlyEmissionChangesByEVType).reduce(
+function calculateTotalMonthlyEmissionChanges(
+  monthlyEmissionChanges: MonthlyEmissionChanges,
+) {
+  const result = Object.entries(monthlyEmissionChanges).reduce(
     (totals, [key, data]) => {
       const month = Number(key);
 
@@ -295,24 +184,20 @@ export function calculateEere({
   regionLineLoss, // region.lineLoss
   regionalLoad, // region.rdf.regional_load
   eereDefaults, // region.eereDefaults.data
-  monthlyVMTPerVehicle, // transportation.monthlyVMTPerVehicle
   dailyStats, // transportation.dailyStats
   hourlyEVChargingPercentages, // transportation.hourlyEVChargingPercentages
-  vehiclesDisplaced, // transportation.vehiclesDisplaced
   monthlyDailyEVEnergyUsage, // transportation.monthlyDailyEVEnergyUsage
-  monthlyEmissionRates, // transportation.monthlyEmissionRates
+  monthlyEmissionChanges, // transportation.monthlyEmissionChanges
   eereTextInputs, // eere.inputs (scaled for each region)
 }: {
   regionMaxEEPercent: number;
   regionLineLoss: number;
   regionalLoad: RegionalLoadData[];
   eereDefaults: EereDefaultData[];
-  monthlyVMTPerVehicle: MonthlyVMTPerVehicle;
   dailyStats: DailyStats;
   hourlyEVChargingPercentages: HourlyEVChargingPercentages;
-  vehiclesDisplaced: VehiclesDisplaced;
   monthlyDailyEVEnergyUsage: MonthlyDailyEVEnergyUsage;
-  monthlyEmissionRates: MonthlyEmissionRates;
+  monthlyEmissionChanges: MonthlyEmissionChanges;
   eereTextInputs: { [field in EereTextInputFieldName]: number };
 }) {
   const {
@@ -344,15 +229,8 @@ export function calculateEere({
   const percentHours = broadProgram ? 100 : topHours;
   const topPercentile = stats.percentile(hourlyLoads, 1 - percentHours / 100);
 
-  const monthlyEmissionChangesByEVType =
-    calculateMonthlyEmissionChangesByEVType({
-      monthlyVMTPerVehicle,
-      vehiclesDisplaced,
-      monthlyEmissionRates,
-    });
-
   const totalMonthlyEmissionChanges = calculateTotalMonthlyEmissionChanges(
-    monthlyEmissionChangesByEVType,
+    monthlyEmissionChanges,
   );
 
   const totalYearlyEmissionChanges = calculateTotalYearlyEmissionChanges(
